+++ test/runtests.jl
@@ -587,12 +587,12 @@ end
     @test @inferred(measurement(" -12.34e1  +-  0.56e1 ")) ≈ -123.4 ± 5.6
     @test @inferred(measurement(" 1234e-1  +/-  5.6e0 ")) ≈   123.4 ± 5.6
     @test @inferred(measurement("  -1234e-1  ")) ≈ measurement(-1234e-1)
-    for a in (w, x, y); @test @inferred(measurement(repr(a))) == a; end
-    @test_throws ErrorException measurement("abc")
-    @test_throws ErrorException measurement("(2±1")
-    @test_throws ErrorException measurement("2±1)e2")
-    @test_throws ErrorException measurement("(2)")
-    @test_throws ErrorException measurement("(2)e-2")
+    for a in (w, x, y); @test @inferred(parse(Measurement{Float64}, repr(a))) == a; end
+    @test_throws ArgumentError measurement("abc")
+    @test_throws ArgumentError measurement("(2±1")
+    @test_throws ArgumentError measurement("2±1)e2")
+    @test_throws ArgumentError measurement("(2)")
+    @test_throws ArgumentError measurement("(2)e-2")
 end
 
 @testset "Fixed bugs" begin